Fixed color conversion to RGB, removed unused code
- Fixed user colors not being converted from HEX to RGB (#11)
- Updated the rayshud.json to work with the latest changes.
- Removed unused code.
